refactor(temporal): move deepEqual and constants to types.ts
Extract deepEqual, retryPolicy, CONTINUE_AS_NEW_THRESHOLD, and
EVENT_BATCH_SIZE from workflows.ts into the shared types module.

Committed-By-Agent: claude
