+++ apps/service/src/temporal/workflows.ts
@@ -9,33 +9,13 @@ import {
 } from '@temporalio/workflow'
 
 import type { SyncActivities } from './activities.js'
+import {
+  deepEqual,
+  CONTINUE_AS_NEW_THRESHOLD,
+  EVENT_BATCH_SIZE,
+  retryPolicy,
+} from './types.js'
 import type { WorkflowStatus } from './types.js'
-
-const CONTINUE_AS_NEW_THRESHOLD = 500
-
-function deepEqual(a: unknown, b: unknown): boolean {
-  if (a === b) return true
-  if (a === null || b === null || typeof a !== 'object' || typeof b !== 'object') return false
-  if (Array.isArray(a) !== Array.isArray(b)) return false
-  if (Array.isArray(a)) {
-    const bArr = b as unknown[]
-    return a.length === bArr.length && a.every((v, i) => deepEqual(v, bArr[i]))
-  }
-  const aKeys = Object.keys(a as object)
-  const bKeys = Object.keys(b as object)
-  if (aKeys.length !== bKeys.length) return false
-  return aKeys.every((k) =>
-    deepEqual((a as Record<string, unknown>)[k], (b as Record<string, unknown>)[k])
-  )
-}
-const EVENT_BATCH_SIZE = 50
-
-const retryPolicy = {
-  initialInterval: '1s',
-  backoffCoefficient: 2.0,
-  maximumInterval: '5m',
-  maximumAttempts: 10,
-} as const
 
 // Setup/teardown: 2m with retry
 const { setup, teardown } = proxyActivities<SyncActivities>({
